requirements.txt
streamlit
pandas
numpy
scikit-learn
xgboost
lightgbm
catboost
shap
optuna
plotly
requests
python-dotenv
chardet
pytest
pyyaml
